·“翔鹤改二→翔鹤改二甲”中断检查修改为“翔鹤改二甲→翔鹤改二” ·启动时预先加载本地api_start2（用于离线查看各种图鉴等功能，实际游戏start2数据一直来自服务器） ·powerup, destoryitem2, destoryship修改为在正确接收服务器结果之后再修改数据集合 ·调整自定义舰队编成的颜色在暗色主题下的异常

+++ ElectronicObserver/Observer/kcsapi/api_req_kaisou/powerup.cs
@@ -9,48 +9,57 @@ namespace ElectronicObserver.Observer.kcsapi.api_req_kaisou {
 
 	public class powerup : APIBase {
 
+		Dictionary<string, string> request;
 
 		public override void OnRequestReceived( Dictionary<string, string> data ) {
-
-			KCDatabase db = KCDatabase.Instance;
-
-
-			db.Fleet.LoadFromRequest( APIName, data );
-
-
-			foreach ( string id in data["api_id_items"].Split( ",".ToCharArray() ) ) {
-
-				int shipID = int.Parse( id );
-
-
-				ShipData ship = db.Ships[shipID];
-				for ( int i = 0; i < ship.Slot.Count; i++ ) {
-					if ( ship.Slot[i] != -1 )
-						db.Equipments.Remove( ship.Slot[i] );
-				}
-
-				Utility.Logger.Add( 2, ship.NameWithLevel + " を除籍しました。" );
-				db.Ships.Remove( shipID );
-
-			}
+			request = data;
 
 			base.OnRequestReceived( data );
 		}
 
 
-		public override void OnResponseReceived( dynamic data ) {
+		public override void OnResponseReceived( dynamic data )
+		{
 
 			KCDatabase db = KCDatabase.Instance;
 
-			var ship = db.Ships[(int)data.api_ship.api_id];
-			if ( ship != null )
-				ship.LoadFromResponse( APIName, data.api_ship );
-
-			db.Fleet.LoadFromResponse( APIName, data.api_deck );
+			if ( request != null )
+			{
 
 
-			if ( Utility.Configuration.Config.Log.ShowSpoiler )
-				Utility.Logger.Add( 2, string.Format( "{0} の近代化改修に{1}しました。", ship.NameWithLevel, ( (int)data.api_powerup_flag ) != 0 ? "成功" : "失敗" ) );
+				db.Fleet.LoadFromRequest( APIName, request );
+
+
+				foreach ( string id in request["api_id_items"].Split( ",".ToCharArray() ) )
+				{
+
+					int shipID = int.Parse( id );
+
+
+					ShipData ship = db.Ships[shipID];
+					for ( int i = 0; i < ship.Slot.Count; i++ )
+					{
+						if ( ship.Slot[i] != -1 )
+							db.Equipments.Remove( ship.Slot[i] );
+					}
+
+					Utility.Logger.Add( 2, ship.NameWithLevel + " を除籍しました。" );
+					db.Ships.Remove( shipID );
+
+				}
+			}
+
+			{
+				var ship = db.Ships[(int)data.api_ship.api_id];
+				if ( ship != null )
+					ship.LoadFromResponse( APIName, data.api_ship );
+
+				db.Fleet.LoadFromResponse( APIName, data.api_deck );
+
+
+				if ( Utility.Configuration.Config.Log.ShowSpoiler )
+					Utility.Logger.Add( 2, string.Format( "{0} の近代化改修に{1}しました。", ship.NameWithLevel, ( (int)data.api_powerup_flag ) != 0 ? "成功" : "失敗" ) );
+			}
 
 			base.OnResponseReceived( (object)data );
 		}
